ANDROID-598 No task history displayed
